reorganizes src/
