+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "extremely-online",
-  "version": "0.1.0",
+  "name": "untitled-webextension-framework",
+  "version": "0.10.0",
   "scripts": {
     "precompile": "rimraf dist && rimraf out-tsc && tsc -p tsconfig.json",
     "build": "npm run precompile && npm run build:options-page && npm run build:addon",
@@ -41,9 +41,7 @@
     "setupFiles": [
       "jest-webextension-mock"
     ],
-    "transformIgnorePatterns": [
-      "/node_modules/@mozilla/rally.js"
-    ]
+    "transformIgnorePatterns": []
   },
   "devDependencies": {
     "@babel/plugin-transform-runtime": "^7.13.10",
@@ -108,14 +106,14 @@
   "description": "A web extension that measures where you go and what you do on the web.",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/mozilla-rally/rally-study-01.git"
+    "url": "git+https://github.com/hamilton/untitled-webextension-framework.git"
   },
   "keywords": [],
-  "author": "Mozilla Rally",
+  "author": "Hamilton Ulmer",
   "license": "MPL-2.0",
   "private": true,
   "bugs": {
-    "url": "https://github.com/mozilla-rally/rally-study-01/issues"
+    "url": "https://github.com/hamilton/untitled-webextension-framework/issues"
   },
-  "homepage": "https://github.com/mozilla-rally/rally-study-01#readme"
+  "homepage": "https://github.com/hamilton/untitled-webextension-framework#readme"
 }
