+++ src/lib/config-interface.ts
