+++ src/lib/event-stream-inspector.ts
