+++ src/lib/event-stream-storage.js
