+++ src/lib/options-page/components/NamespaceControls.svelte
