+++ src/lib/options-page/components/NamespaceGrid.svelte
