+++ src/lib/options-page/global.d.ts
