+++ src/lib/options-page/main.ts
@@ -1,5 +1,5 @@
 import App from './routes/App.svelte';
-import { modules } from '../../app.config';
+import { modules } from '../../../app.config';
 
 const app = new App({
 	target: document.body,
