--- src/main.js
@@ -1,4 +1,0 @@
-// import { Rally, runStates } from "@mozilla/rally";
-import run from "../lib/rallytv"
-
-run();
